Add test runner
This acts as an entry point for running the tests. To run them, call:

    python forkit/tests/run.py

This runs the tests against a Postgres database. This can be overridden
with a DATABASE_URL environment variable, as documented on the
django-environ website [0].

If you use the default DB settings, you'll need to create a database
before running the tests:

    psql -c "CREATE DATABASE forkit;"

To get the new Django test runner to discover the existing tests, I had
to rename the files so that they started with `test_`.

This also removes the custom test runner, colourer, and coverage module.
I'll restore this functionality in another commit.

[0]: https://django-environ.readthedocs.io/en/latest/