Fork it; we're forking forkit
I've had to duplicate the BSD licence, so that we can licence our new
changes without the modified 3rd clause of the original, while still
maintaining the historical licence on the old code.

I'm not a lawyer, but I think that's how to deal with this? Please
open an issue on this repo if you think I've made a mistake here.

https://github.com/Virtualstock/forkit-django